requirements.txt
twilio==7.13.0
telethon>=1.25.0
python-dotenv==0.21.0